[AppSec] Merge DataContract extraction into ObjectExtractor
