+++ tracer/src/Datadog.Trace/AppSec/ObjectExtractor.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="ObjectExtractor.cs" company="Datadog">
+// <copyright file="ObjectExtractor.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -11,12 +11,11 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
+using System.Runtime.Serialization;
 using Datadog.Trace.AppSec.Waf;
 using Datadog.Trace.Logging;
 using Datadog.Trace.Util;
 using Datadog.Trace.Vendors.Serilog.Events;
-
-#nullable enable
 
 namespace Datadog.Trace.AppSec
 {
@@ -25,7 +24,8 @@ namespace Datadog.Trace.AppSec
         private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor(typeof(ObjectExtractor));
         private static readonly IReadOnlyDictionary<string, object?> EmptyDictionary = new ReadOnlyDictionary<string, object?>(new Dictionary<string, object?>(0));
 
-        private static readonly ConcurrentDictionary<Type, FieldExtractor?[]> TypeToExtractorMap = new();
+        private static readonly ConcurrentDictionary<Type, MemberExtractor?[]> TypeToExtractorMap = new();
+        private static readonly ConcurrentDictionary<Type, MemberExtractor?[]> DataContractTypeToExtractorMap = new();
 
         private static readonly HashSet<Type> WafProcessableTypes =
         [
@@ -43,18 +43,32 @@ namespace Datadog.Trace.AppSec
         ];
 
         internal static object? Extract(object? body)
+            => Extract(body, TypeToExtractorMap, CreateDefaultExtractors);
+
+        internal static object? ExtractDataContract(object? body)
+            => Extract(body, DataContractTypeToExtractorMap, CreateDataContractAwareExtractors);
+
+        private static object? Extract(
+            object? body,
+            ConcurrentDictionary<Type, MemberExtractor?[]> extractorCache,
+            Func<Type, MemberExtractor?[]> createExtractors)
         {
-            if (body == null)
+            if (body is null)
             {
                 return null;
             }
 
             var visited = new HashSet<object>();
-            var item = ExtractType(body.GetType(), body, 0, visited);
+            var item = ExtractType(body.GetType(), body, 0, visited, extractorCache, createExtractors);
             return item;
         }
 
-        private static IReadOnlyDictionary<string, object?> ExtractProperties(object body, int depth, HashSet<object> visited)
+        private static IReadOnlyDictionary<string, object?> ExtractProperties(
+            object body,
+            int depth,
+            HashSet<object> visited,
+            ConcurrentDictionary<Type, MemberExtractor?[]> extractorCache,
+            Func<Type, MemberExtractor?[]> createExtractors)
         {
             try
             {
@@ -79,83 +93,198 @@ namespace Datadog.Trace.AppSec
 
             depth++;
 
-            if (!TypeToExtractorMap.TryGetValue(bodyType, out var fieldExtractors))
+            if (!extractorCache.TryGetValue(bodyType, out var memberExtractors))
             {
-                var fields = bodyType.GetFields(BindingFlags.Instance | BindingFlags.NonPublic);
-                if (!bodyType.GetTypeInfo().IsAnonymous())
-                {
-                    fields = fields.Where(x => x.IsPrivate && x.Name.EndsWith("__BackingField")).ToArray();
-                }
-
-                fieldExtractors = new FieldExtractor[fields.Length];
-                for (var i = 0; i < fields.Length; i++)
-                {
-                    var field = fields[i];
-
-                    var propertyName = GetPropertyName(field.Name);
-                    if (StringUtil.IsNullOrEmpty(propertyName))
-                    {
-                        Log.Warning("ExtractProperties - couldn't extract property name from: {FieldName}", field.Name);
-                        continue;
-                    }
-
-                    var dynMethod = new DynamicMethod(
-                        bodyType + "_get_" + propertyName,
-                        typeof(object),
-                        [typeof(object)],
-                        typeof(ObjectExtractor).Module,
-                        true);
-                    var ilGen = dynMethod.GetILGenerator();
-                    ilGen.Emit(OpCodes.Ldarg_0);
-                    if (bodyType.IsValueType)
-                    {
-                        ilGen.Emit(OpCodes.Unbox_Any, bodyType);
-                    }
-
-                    ilGen.Emit(OpCodes.Ldfld, field);
-                    if (field.FieldType.IsValueType)
-                    {
-                        ilGen.Emit(OpCodes.Box, field.FieldType);
-                    }
-
-                    ilGen.Emit(OpCodes.Ret);
-                    var func = (Func<object, object>)dynMethod.CreateDelegate(typeof(Func<object, object>));
-
-                    fieldExtractors[i] = new FieldExtractor(propertyName!, field.FieldType, func);
-                }
+                memberExtractors = createExtractors(bodyType);
 
                 // this would be expected to fail sometimes, when several threads attempt process the same body
-                TypeToExtractorMap.TryAdd(bodyType, fieldExtractors);
+                extractorCache.TryAdd(bodyType, memberExtractors);
             }
 
-            var dictSize = Math.Min(WafConstants.MaxContainerSize, fieldExtractors.Length);
+            var dictSize = Math.Min(WafConstants.MaxContainerSize, memberExtractors.Length);
             var dict = new Dictionary<string, object?>(dictSize);
 
-            for (var i = 0; i < fieldExtractors.Length; i++)
+            for (var i = 0; i < memberExtractors.Length; i++)
             {
                 if (dict.Count >= WafConstants.MaxContainerSize || depth >= WafConstants.MaxContainerDepth)
                 {
                     return dict;
                 }
 
-                var fieldExtractor = fieldExtractors[i];
+                var memberExtractor = memberExtractors[i];
 
-                if (fieldExtractor is not null)
+                if (memberExtractor is not null)
                 {
-                    var value = fieldExtractor.Accessor.Invoke(body);
+                    var value = memberExtractor.Accessor.Invoke(body);
                     if (Log.IsEnabled(LogEventLevel.Debug))
                     {
-                        Log.Debug("ExtractProperties - property: {BodyType}.{Name} {Value}", bodyType.FullName, fieldExtractor.Name, value);
+                        Log.Debug("ExtractProperties - property: {BodyType}.{Name} {Value}", bodyType.FullName, memberExtractor.Name, value);
                     }
 
-                    var item =
-                        value == null ? null : ExtractType(fieldExtractor.Type, value, depth, visited);
+                    var item = value is null ? null : ExtractType(memberExtractor.Type, value, depth, visited, extractorCache, createExtractors);
 
-                    dict.Add(fieldExtractor.Name, item);
+                    dict.Add(memberExtractor.Name, item);
                 }
             }
 
             return dict;
+        }
+
+        private static MemberExtractor?[] CreateDefaultExtractors(Type bodyType)
+        {
+            var fields = bodyType.GetFields(BindingFlags.Instance | BindingFlags.NonPublic);
+            if (!bodyType.GetTypeInfo().IsAnonymous())
+            {
+                fields = fields.Where(x => x.IsPrivate && x.Name.EndsWith("__BackingField")).ToArray();
+            }
+
+            var fieldExtractors = new MemberExtractor[fields.Length];
+            for (var i = 0; i < fields.Length; i++)
+            {
+                var field = fields[i];
+
+                var propertyName = GetPropertyName(field.Name);
+                if (StringUtil.IsNullOrEmpty(propertyName))
+                {
+                    Log.Warning("ExtractProperties - couldn't extract property name from: {FieldName}", field.Name);
+                    continue;
+                }
+
+                fieldExtractors[i] = new MemberExtractor(propertyName!, field.FieldType, CreateFieldAccessor(bodyType, field));
+            }
+
+            return fieldExtractors;
+        }
+
+        private static MemberExtractor?[] CreateDataContractAwareExtractors(Type bodyType)
+        {
+            return bodyType.GetCustomAttribute<DataContractAttribute>() is not null
+                ? CreateDataContractExtractors(bodyType)
+                : CreateDataContractDefaultExtractors(bodyType);
+        }
+
+        private static MemberExtractor?[] CreateDataContractExtractors(Type bodyType)
+        {
+            var extractors = new List<MemberExtractor>();
+            foreach (var property in bodyType.GetProperties(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic))
+            {
+                if (property.GetIndexParameters().Length > 0
+                    || property.GetCustomAttribute<IgnoreDataMemberAttribute>() is not null
+                    || property.GetCustomAttribute<DataMemberAttribute>() is not { } dataMember
+                    || property.GetGetMethod(nonPublic: true) is not { } getter)
+                {
+                    continue;
+                }
+
+                var name = StringUtil.IsNullOrEmpty(dataMember.Name) ? property.Name : dataMember.Name;
+                extractors.Add(new MemberExtractor(name!, property.PropertyType, CreatePropertyAccessor(bodyType, getter, property.PropertyType)));
+            }
+
+            foreach (var field in bodyType.GetFields(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic))
+            {
+                if (field.GetCustomAttribute<IgnoreDataMemberAttribute>() is not null
+                    || field.GetCustomAttribute<DataMemberAttribute>() is not { } dataMember)
+                {
+                    continue;
+                }
+
+                var name = StringUtil.IsNullOrEmpty(dataMember.Name) ? GetPropertyName(field.Name) ?? field.Name : dataMember.Name;
+                extractors.Add(new MemberExtractor(name!, field.FieldType, CreateFieldAccessor(bodyType, field)));
+            }
+
+            return extractors.ToArray();
+        }
+
+        private static MemberExtractor?[] CreateDataContractDefaultExtractors(Type bodyType)
+        {
+            var fields = bodyType.GetFields(BindingFlags.Instance | BindingFlags.NonPublic);
+            if (!bodyType.GetTypeInfo().IsAnonymous())
+            {
+                fields = fields.Where(x => x.IsPrivate && x.Name.EndsWith("__BackingField")).ToArray();
+            }
+
+            var fieldExtractors = new MemberExtractor[fields.Length];
+            for (var i = 0; i < fields.Length; i++)
+            {
+                var field = fields[i];
+                var propertyName = GetPropertyName(field.Name);
+                if (StringUtil.IsNullOrEmpty(propertyName))
+                {
+                    Log.Warning("ExtractProperties - couldn't extract property name from: {FieldName}", field.Name);
+                    continue;
+                }
+
+                var property = bodyType.GetProperty(propertyName!, BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
+                if (property?.GetCustomAttribute<IgnoreDataMemberAttribute>() is not null)
+                {
+                    continue;
+                }
+
+                var dataMemberName = property?.GetCustomAttribute<DataMemberAttribute>()?.Name;
+                var name = StringUtil.IsNullOrEmpty(dataMemberName) ? propertyName : dataMemberName;
+                fieldExtractors[i] = new MemberExtractor(name!, field.FieldType, CreateFieldAccessor(bodyType, field));
+            }
+
+            return fieldExtractors;
+        }
+
+        private static Func<object, object?> CreateFieldAccessor(Type bodyType, FieldInfo field)
+        {
+            var dynMethod = new DynamicMethod(
+                bodyType + "_get_" + field.Name,
+                typeof(object),
+                [typeof(object)],
+                typeof(ObjectExtractor).Module,
+                true);
+            var ilGen = dynMethod.GetILGenerator();
+            ilGen.Emit(OpCodes.Ldarg_0);
+            if (bodyType.IsValueType)
+            {
+                ilGen.Emit(OpCodes.Unbox_Any, bodyType);
+            }
+            else
+            {
+                ilGen.Emit(OpCodes.Castclass, bodyType);
+            }
+
+            ilGen.Emit(OpCodes.Ldfld, field);
+            if (field.FieldType.IsValueType)
+            {
+                ilGen.Emit(OpCodes.Box, field.FieldType);
+            }
+
+            ilGen.Emit(OpCodes.Ret);
+            return (Func<object, object?>)dynMethod.CreateDelegate(typeof(Func<object, object?>));
+        }
+
+        private static Func<object, object?> CreatePropertyAccessor(Type bodyType, MethodInfo getter, Type propertyType)
+        {
+            var dynMethod = new DynamicMethod(
+                bodyType + "_get_" + getter.Name,
+                typeof(object),
+                [typeof(object)],
+                typeof(ObjectExtractor).Module,
+                true);
+            var ilGen = dynMethod.GetILGenerator();
+            ilGen.Emit(OpCodes.Ldarg_0);
+            if (bodyType.IsValueType)
+            {
+                ilGen.Emit(OpCodes.Unbox, bodyType);
+                ilGen.Emit(OpCodes.Call, getter);
+            }
+            else
+            {
+                ilGen.Emit(OpCodes.Castclass, bodyType);
+                ilGen.Emit(getter.IsVirtual ? OpCodes.Callvirt : OpCodes.Call, getter);
+            }
+
+            if (propertyType.IsValueType)
+            {
+                ilGen.Emit(OpCodes.Box, propertyType);
+            }
+
+            ilGen.Emit(OpCodes.Ret);
+            return (Func<object, object?>)dynMethod.CreateDelegate(typeof(Func<object, object?>));
         }
 
         private static string? GetPropertyName(string fieldName)
@@ -170,7 +299,13 @@ namespace Datadog.Trace.AppSec
             return null;
         }
 
-        private static object? ExtractType(Type itemType, object value, int depth, HashSet<object> visited)
+        private static object? ExtractType(
+            Type itemType,
+            object value,
+            int depth,
+            HashSet<object> visited,
+            ConcurrentDictionary<Type, MemberExtractor?[]> extractorCache,
+            Func<Type, MemberExtractor?[]> createExtractors)
         {
             if (itemType == typeof(string))
             {
@@ -179,19 +314,19 @@ namespace Datadog.Trace.AppSec
 
             if (itemType.IsArray || (itemType.IsGenericType && itemType.GetGenericTypeDefinition() == typeof(List<>)))
             {
-                return ExtractListOrArray(value, depth, visited);
+                return ExtractListOrArray(value, depth, visited, extractorCache, createExtractors);
             }
 
             if (itemType.IsGenericType && itemType.GetGenericTypeDefinition() == typeof(Dictionary<,>))
             {
-                return ExtractDictionary(value, itemType, depth, visited);
+                return ExtractDictionary(value, itemType, depth, visited, extractorCache, createExtractors);
             }
 
             // case of System.Web.Routing.RouteValueDictionary and types inheriting IDictionary
             var iDictionaryType = typeof(IDictionary<string, object>);
             if (iDictionaryType.IsAssignableFrom(itemType))
             {
-                return ExtractDictionary(value, iDictionaryType, depth, visited);
+                return ExtractDictionary(value, iDictionaryType, depth, visited, extractorCache, createExtractors);
             }
 
             if (WafProcessableTypes.Contains(itemType))
@@ -208,11 +343,17 @@ namespace Datadog.Trace.AppSec
                 return value.ToString();
             }
 
-            var nestedDict = ExtractProperties(value, depth, visited);
+            var nestedDict = ExtractProperties(value, depth, visited, extractorCache, createExtractors);
             return nestedDict;
         }
 
-        private static IReadOnlyDictionary<string, object?> ExtractDictionary(object value, Type dictType, int depth, HashSet<object> visited)
+        private static IReadOnlyDictionary<string, object?> ExtractDictionary(
+            object value,
+            Type dictType,
+            int depth,
+            HashSet<object> visited,
+            ConcurrentDictionary<Type, MemberExtractor?[]> extractorCache,
+            Func<Type, MemberExtractor?[]> createExtractors)
         {
             if (!visited.Add(value))
             {
@@ -253,7 +394,7 @@ namespace Datadog.Trace.AppSec
                         }
                         else
                         {
-                            var extractedvalue = ExtractType(dictValue.GetType(), dictValue, depth + 1, visited);
+                            var extractedvalue = ExtractType(dictValue.GetType(), dictValue, depth + 1, visited, extractorCache, createExtractors);
                             items.Add(dictKey, extractedvalue);
                         }
 
@@ -268,7 +409,12 @@ namespace Datadog.Trace.AppSec
             return items;
         }
 
-        private static List<object?> ExtractListOrArray(object value, int depth, HashSet<object> visited)
+        private static List<object?> ExtractListOrArray(
+            object value,
+            int depth,
+            HashSet<object> visited,
+            ConcurrentDictionary<Type, MemberExtractor?[]> extractorCache,
+            Func<Type, MemberExtractor?[]> createExtractors)
         {
             if (visited.Contains(value))
             {
@@ -287,7 +433,7 @@ namespace Datadog.Trace.AppSec
                 }
                 else
                 {
-                    var extractedvalue = ExtractType(item.GetType(), item, depth + 1, visited);
+                    var extractedvalue = ExtractType(item.GetType(), item, depth + 1, visited, extractorCache, createExtractors);
                     items.Add(extractedvalue);
                 }
 
@@ -300,13 +446,13 @@ namespace Datadog.Trace.AppSec
             return items;
         }
 
-        private sealed class FieldExtractor(string name, Type type, Func<object, object> accessor)
+        private sealed class MemberExtractor(string name, Type type, Func<object, object?> accessor)
         {
             public string Name { get; } = name;
 
             public Type Type { get; } = type;
 
-            public Func<object, object> Accessor { get; } = accessor;
+            public Func<object, object?> Accessor { get; } = accessor;
         }
     }
 }
